halucinator/pyproject.toml
[tool.poetry]
name = "halucinator"
version = "0.1.0"
description = ""
authors = ["Antony Vennard <antony@vennard.ch>"]

[tool.poetry.scripts]
halucinator = "halucinator.main:main"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML ="^6.0"
pyzmq = "20.0.0"
IPython = "^8.5.0"
pyelftools = "^0.29"
keystone-engine = "^0.9.2"
nose = "^1.3.7"
cxxfilt = "^0.3.0"
angr = "^9.2.22"
ipdb = "^0.13.9"
python-prctl = "^1.8.1"
docker = "6.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pdoc3 = "0.10.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
